+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-17T02:38:11.740Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-17T13:05:36.548Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")