+++ packages/orq-rc/RELEASES.md
@@ -11029,3 +11029,13 @@ Based on:
 - [python v4.16.0-rc.0] packages/orq-rc
 ### Releases
 - [PyPI v4.16.0-rc.0] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.0 - packages/orq-rc
+
+## 2026-09-17 13:08:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.797.0 (2.937.18) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.16.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v4.16.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.1 - packages/orq-rc